requirements.txt
aniso8601==9.0.1
biopython==1.79
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
Cython==0.29.32
Flask==2.2.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mock==4.0.3
numpy==1.23.4
pycodestyle==2.9.1
pytz==2022.4
requests==2.28
six==1.16.0
requests-to-curl==1.1.0
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
python-dotenv==1.0.0
